backend/requirements.txt
aniso8601==9.0.1
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-restx==1.1.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
